Exit + Enter ship. Landing adjustment
Added the ability to exit the ship and move around as the player and the ability to re-enter the ship to take off from the landing pad.

Adjusted the align rotation portion of the landing sequence to 0 out pitch and roll and not descend early. Also adjusted rotation logic so the ship will properly rotate itself regardless of the landing pads yaw